Base reading/saving of configuration file on program invocation name.
Top from procps saves a config file based on invocation name. If called as top, it saves a ~/.toprc file, if called as bottom (say, using a symlink) it saves a ~/.bottomrc. This way one can have a number of different setups depending on how you call the same program.