Modify proc read routines to accomodate big SMP systems, also minor formatting fix for load average display.
Change the operation to read another chunk of the file in case memchr
can not find a newline in the current buffer. This only happens on very
big SMP systems, where lines can get very long due to a large number of
cores (in this case the "intr" line in /proc/stat which caused the
read_btime function to exit bevore having read the btime). Also modify
the load average display format to accomodate load numbers above 1000
without growing to wide or shrinking on lower loads.